Refactor oesign handling of .oeinfo
- Refactor out `_sgx_load_enclave_properties()` as shared `oe_read_oeinfo_sgx()`
  into new oeinfo.c file.
- Refactor out `_update_and_write_signed_exe()` as shared `oe_write_oeinfo_sgx()`
  into new oeinfo.c file.
- Replace `oe_sgx_load_properties()` and `_find_enclave_properties()` used in
  oedump.c with `oe_read_oeinfo_sgx()`.
