+++ tools/oesign/CMakeLists.txt
@@ -6,7 +6,8 @@ include(CheckSymbolExists)
 list(APPEND SOURCES
   main.c
   oe_err.c
-  oedump.c)
+  oedump.c
+  oeinfo.c)
 
 CHECK_SYMBOL_EXISTS(getopt_long getopt.h HAVE_GETOPT_LONG)
